+++ src/app/(header)/explore/_components/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useTab } from './useTab';
